Store multiple mod names to support fuzzy search

+++ GitGudModsListLoader/Models/ModShortInfo.cs
@@ -1,3 +1,3 @@
 ﻿namespace GitGudModsListLoader.Models;
 
-public record ModShortInfo(int Id, string Title, int ProjectId, string MetadataPath = "src/meta.ini");
+public record ModShortInfo(int Id, string Title, long ProjectId, string MetadataPath = "src/meta.ini");
